Refactoring, error and info messages
Due to the distributed artchecture of the OneView Daemon and Ironic,
ironic-oneviewd may think a node is in a different provision state.
The code already shows a message when this occurs. However, the
message is a traceback and may lead the user to think the process is
broken. This commit changes the error message to the default message
instead of the traceback. It also removes the need to take nodes
that are available.